Add Script(Boss Object Setup)

+++ PC/Project-FW/Bounding.h
@@ -6,18 +6,18 @@ typedef struct _IBounding
 
 typedef struct _BoundingBox : public Bounding
 {
-	int _left, _top ;
-	int _right, _bottom ;
+	float _left, _top ;
+	float _right, _bottom ;
 
-	_BoundingBox(int left=0, int top=0, int right=0, int bottom=0) : _left(left), _top(top), _right(right), _bottom(bottom)
+	_BoundingBox(float left=0.0f, float top=0.0f, float right=0.0f, float bottom=0.0f) : _left(left), _top(top), _right(right), _bottom(bottom)
 	{}
 } BoundingBox, BBox ;
 
 typedef struct _BoundingCircle : public Bounding
 {
-	int _x, _y ;
+	float _x, _y ;
 	float _radius ;
 
-	_BoundingCircle(int x=0, int y=0, float radius=0.0f) : _x(x), _y(y), _radius(radius)
+	_BoundingCircle(float x=0.0f, float y=0.0f, float radius=0.0f) : _x(x), _y(y), _radius(radius)
 	{}
 } BoundingCircle, BCircle ;
